boulder-oracle-sprint: Merged to [5306]
git-svn-id: http://code.djangoproject.com/svn/django/branches/boulder-oracle-sprint@5307 bcc190cf-cafb-0310-a4f2-bffc1f526a37

+++ django/db/backends/oracle/introspection.py
@@ -92,7 +92,7 @@ DATA_TYPES_REVERSE = {
     cx_Oracle.DATETIME: 'DateTimeField',
     cx_Oracle.FIXED_CHAR: 'CharField',
     cx_Oracle.NCLOB: 'TextField',
-    cx_Oracle.NUMBER: 'FloatField',
+    cx_Oracle.NUMBER: 'DecimalField',
     cx_Oracle.STRING: 'CharField',
     cx_Oracle.TIMESTAMP: 'DateTimeField',
 }
